my-react-cdk-python-app/requirements.txt
Flask==2.0.1
aws-cdk.core==1.100.0
aws-cdk.aws-s3==1.100.0
aws-cdk.aws-route53==1.100.0
aws-cdk.aws-s3-deployment==1.100.0
aws-cdk.aws-certificatemanager==1.100.0
boto3==1.18.0